- all: version 0.28 - cantomat: rewrote signal name prefix code (replaced -f by -p) - libmatfile: new library for MAT file handling

+++ src/cantomat/busassignment.h
@@ -22,7 +22,7 @@
 #include "messagehash.h"
 
 typedef struct {
-  int bus;
+  uint16_t bus;
   char *filename;
   messageHash_t *messageHash;
 } busAssignmentEntry_t;
